pyproject.toml
[project]
name = "outlier-bench"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "numpy>=2.4.0",
    "pandas>=2.3.3",
    "pydantic>=2.12.5",
    "pyyaml>=6.0.3",
    "scikit-learn>=1.8.0",
    "tabulate>=0.9.0",
    "typer>=0.21.1",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.14.10",
]

[project.scripts]
outlier-bench = "outlier_bench.cli:app"

[tool.ruff]
line-length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
package = true

[tool.hatch.build.targets.wheel]
packages = ["src/outlier_bench"]

